backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
pandas
flasgger
flask_cors
pytest
marshmallow==3.13.0